Relocate some components to their respecive folders in the screens folder.
