+++ screens/pokemon/tabs/about/components/breeding/GenderBalanceBar.tsx
