+++ screens/pokemon/tabs/about/components/description/PokemonDescription.tsx
@@ -2,15 +2,16 @@ import { StyleSheet, View } from 'react-native'
 
 import { CustomText } from '@/components/typography/customText'
 import { typography } from '@/constants/typography'
-
-import { TabSectionHeader } from '@/components/headers/TabSectionHeader'
-import palette from '@/constants/palette'
 import { formatText } from '@/utils/formatters/formatTextString'
+
+import { PokeTabSectionHeader } from '../../../PokeTabSectionHeader'
+
+import palette from '@/constants/palette'
 
 export const PokemonDescription = ({ description }: { description: string }) => {
   return (
     <View>
-      <TabSectionHeader title={typography.tabs.description} />
+      <PokeTabSectionHeader title={typography.tabs.description} />
       <CustomText style={styles.descriptionText} numberOfLines={3}>
         {formatText(description)}
       </CustomText>
