+++ screens/pokemons/components/card/PokemonCardImage.tsx
