+++ screens/pokemons/components/card/pokemonCard.tsx
